node: unexport built-in API handler objects
